--- MatematikHelper/solver.py
@@ -1,4 +1,0 @@
-from sympy import *
-# a(3t-1,12)
-# b(6,t+8)
-
