+++ old code/oldcode.py
@@ -1,0 +1,51 @@
+# a(3t-1,12)
+# b(6,t+8)
+# opgaver.opgave1()
+"""
+print("__________Example______________")
+a = V.Vector2D(1,6)
+b = V.Vector2D(5,2)
+c = a.projektering(b)
+print("X:{}\nY:{}".format(c.x,c.y))
+
+print("____________Opgave 1.26____________")
+a = V.Vector2D(1,2)
+b = V.Vector2D(4,1)
+c = a.projektering(b)
+print("a)\nX:{}\nY:{}\n".format(c.x,c.y))
+
+a = V.Vector2D(-3,-1)
+b = V.Vector2D(-5,4)
+c = a.projektering(b)
+print("b)\nX:{}\nY:{}\n".format(c.x,c.y))
+"""
+"""
+a = V.Vector2D(-6,3)
+b = V.Vector2D(10,4)
+c = V.Vector2D(1,10)
+d = V.Vector2D(20,5)
+L = [a,b,c,d]
+
+V.show_vectors(L)
+"""
+
+"""
+l = GF.AndenGrads("-3*x**2+6*x+2")
+l.differentabel()
+"""
+
+"""
+b = a.tver_vektor()
+print("Vx:{}  Vy:{}".format(b.x,b.y))
+"""
+
+# opgaver.opgave2()
+# x = Symbol("x")
+
+"""
+f = Fun.Function("20+150*log(8*x+1)")
+f.show_function()
+
+#f.show_function(range(-100,100))
+
+"""
