+++ solver.py
@@ -1,0 +1,4 @@
+from sympy import *
+# a(3t-1,12)
+# b(6,t+8)
+
